requirements.txt
Django==4.2.7
mysqlclient==2.2.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
mitmproxy==10.0.0
redis==5.0.0
python-nmap==0.7.1
aiohttp==3.8.5
websockets==11.0.3
wsproto==1.2.0